Separate DayTimeLocation into its own component.
